Move TestComet inside LiftSessionSpec.scala and create it with the newer form of findOrCreateComet.
