+++ web/webkit/src/test/scala/net/liftweb/http/LiftSessionSpec.scala
@@ -17,27 +17,34 @@
 package net.liftweb
 package http
 
+import scala.xml.NodeSeq
 import net.liftweb.common.{Loggable, Full, Empty}
 import org.specs2.mutable.Specification
 
-import scala.xml.NodeSeq
-
 object LiftSessionSpec extends Specification with Loggable {
 
-  val testSession = new LiftSession("Test Session", "", Empty) {
-    override private[liftweb] def set[T](name: String, value: T) = {
-      logger.info(s"set called. $name $value")
-      super.set(name, value)
+  class TestComet extends CometActor with Loggable {
+    logger.info("TestComet created")
+    def render = NodeSeq.Empty
+    override def !(msg: Any) = {
+      logger.info("Message received: " + msg)
     }
   }
 
-
   "A LiftSession" should {
-    "Render messages in the order collected" in {
+
+    "Send accumulated messages to a newly-created comet actor in the order in which they arrived" in {
+
+      val testSession = new LiftSession("Test Session", "", Empty) {
+        override private[liftweb] def set[T](name: String, value: T) = {
+          logger.info(s"set called. $name $value")
+          super.set(name, value)
+        }
+      }
+
       val cometName = "TestComet"
       1 to 3 foreach(n => testSession.sendCometActorMessage("CometType", Full(cometName), n))
-      logger.info(testSession.nmyVariables)
-      val bc = testSession.findOrCreateComet(cometName)
+      val bc = testSession.findOrCreateComet[TestComet](Empty, NodeSeq.Empty, Map.empty)
       logger.info("Comet: " + bc)
       bc.foreach(_.!(1))
       bc.isDefined mustEqual true
